requirements.txt
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
colorama==0.4.6
Django==2.2.16
django-debug-toolbar==3.2.4
Faker==12.0.1
idna==3.7
iniconfig==2.0.0
mixer==7.1.2
packaging==24.0
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.5.0
toml==0.10.2
urllib3==1.26.18
